feature/fix-sonar-issues: Fixing sonar issues + some tests were in the wrong file.
